+++ src/util.cpp
@@ -13,7 +13,7 @@ double toRadians(const double degrees) { return degrees * (M_PI / 180.0); }
 
 
 // Scales a value from a current maximum range to a new maximum range
-size_t scaleValue(const double x, const size_t currMax, const size_t newMax) {
+int scaleValue(const double x, const int currMax, const int newMax) {
   if (!currMax) {
     throw std::invalid_argument("currMax cannot be zero");
   }
